+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 All notable changes will be documented here.
 
+## [0.1.5] - 2022-05-03
+### Added
+- Info object can be replaced with the cacher prop on ``NOAAStormEvents`` or via ``Info.replace``.
+- ``NOAAStormEvents`` now takes an object as the first prop. ``onlyNew`` has been rolled into the prop.
+
 ## [0.1.4] - 2022-02-18
-### Addded
+### Added
 - Info function for saving a history of downloaded files.
 - Ability to set the cache path to a user defined location.
 - Added ``NOAAStormEvents.onlyNew`` to only parse new files.